tianyancha_com/requirements.txt
amqp==1.4.9
anyjson==0.3.3
apscheduler==3.5.3
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
billiard==3.3.0.23
bleach==3.1.0
blinker==1.4
brotlipy==0.7.0
bs4==0.0.1
cchardet==2.1.4
celery==3.1.26.post2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.6.1
colorama==0.4.0
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
dask==0.20.1
decorator==4.3.0
defusedxml==0.5.0
django==2.1.5
django-apscheduler==0.2.13
django-celery==3.2.2
django-cors-headers==2.4.0
entrypoints==0.3
furl==2.0.0
gerapy==0.8.6b2
gevent==1.3.7
greenlet==0.4.15
h11==0.7.0
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kaitaistruct==0.8
kiwisolver==1.0.1
kombu==3.0.37
ldap3==2.5.1
lxml==4.2.5
MarkupSafe==1.1.0
matplotlib==3.0.1
mistune==0.8.4
mitmproxy==4.0.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.15.4+mkl
opencv-python==3.4.3.18
orderedmultidict==1.0
pandas==0.23.4
pandocfilters==1.4.2
paramiko==2.4.2
parsel==1.5.1
parso==0.3.1
passlib==1.7.1
pickleshare==0.7.5
pika==0.12.0
Pillow==5.3.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAutoGUI==0.9.38
pycparser==2.19
PyDispatcher==2.0.5
pydivert==2.1.0
pyexecjs==1.5.1
Pygments==2.2.0
PyHamcrest==1.9.0
pymongo==3.7.2
PyMsgBox==1.0.6
pymysql==0.9.3
PyNaCl==1.3.0
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyquery==1.4.0
PyScreeze==0.1.19
python-dateutil==2.7.5
python-decouple==3.1
python-scrapyd-api==2.1.2
PyTweening==1.0.3
pytz==2018.7
PyWavelets==1.0.1
pywin32==223
pywinpty==0.5.5
pyzmq==17.1.2
qtconsole==4.4.3
queuelib==1.5.0
redis==3.0.1
requests==2.21.0
requests-toolbelt==0.8.0
retrying==1.3.3
ruamel.yaml==0.15.77
scikit-image==0.14.1
scipy==1.1.0
Scrapy==1.5.1
scrapy-redis==0.6.8
scrapy-splash==0.7.2
selenium==3.141.0
Send2Trash==1.5.0
service-identity==17.0.0
six==1.11.0
sortedcontainers==2.0.5
soupsieve==1.8
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
Twisted==18.7.0
tzlocal==2.0.0b1
urllib3==1.24.1
urwid==2.0.1
vine==1.2.0
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wsproto==0.11.0
xlrd==1.1.0
XlsxWriter==1.1.2
zope.interface==4.6.0
